admin/requirements.txt
Flask==3.0.3
flask-restx==1.3.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
pytest==8.3.3
pytest-cov==5.0.0
mock==5.1.0
pytest-mock==3.14.0
pytest-flask==1.3.0
